refactor: 컴포넌트명 변경 및 store 함수 이동
